// src/models/user.ts
export class User{
  achternaam: string;
  achtervang: string;
  created_at: Date;
  id: number;
  kamer: string;
  organisatieonderdeel: string;
  overmij: string;
  rcid: string;
  taken: string;
  telefoonnummermobiel: string;
  telefoonnummervast: string;
  updated_at: Date;
  voornaam: string;
  werktijden: string;
  functie: string;
}